axf/requirement.txt
django==2.0.7
pymysql
djangorestframework==3.9.3
django-filter
django-redis
redis==2.10.6
django-cors-headers==2.5.3